CSS: Improve ShowcaseArticle styling and fix contenteditable outline
The ShowcaseArticle module CSS file has been modified to improve the styling of the article showcase. The issue where it did not look good when there was an <h1> element in the middle of texts has been addressed. Additionally, the outline of the contenteditable element has been set to be transparent.