Add InlineScript component
This commit adds a new file `InlineScript.tsx` to the `src/components` directory. The `InlineScript` component is defined with props `id` and `children`. Inside the component, there is a check for `isExported` using the `useContext` hook. If `isExported` is `false`, it renders a `Script` component with `id` and `children` props. Otherwise, it renders a `script` tag with the `children` prop.